Conflicts:
	arch/avr32/mach-at32ap/include/mach/cpu.h

+++ arch/ia64/include/asm/spinlock_types.h
@@ -7,7 +7,7 @@ FASTVC-BENCH-CORPUS:linux-main-100k-v1-e0bd41dc-8e12-4f1b-978d-a3645ae59da0
 #endif
 
 typedef struct {
-	volatile unsigned long lock;
+	volatile unsigned int lock;
 } raw_spinlock_t;
 
 #define __RAW_SPIN_LOCK_UNLOCKED	{ 0 }
